backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
openai==0.28.1
easyocr==1.7.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
Pillow==10.0.0
matplotlib==3.7.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
python-dotenv==1.0.0
pydantic==2.3.0
gunicorn==21.2.0
pytesseract==0.3.10
pymupdf